feat:add basic classic game, without AI,add drag&drop,add random locations for computer
